Explicitly convert _lh_array_tag_type_value from unsigned to int
Build failed on Mac with Xcode4.6. This is caused by the backport of
JDK-8152856. The definition of _lh_array_tag_type_value changed from int
to unsigned int, causing _lh_array_tag_obj_value being implicitly converted
to unsigned int during the compilation.

+++ src/hotspot/src/share/vm/oops/klass.cpp
@@ -215,7 +215,7 @@ jint Klass::array_layout_helper(BasicType etype) {
   int  hsize = arrayOopDesc::base_offset_in_bytes(etype);
   int  esize = type2aelembytes(etype);
   bool isobj = (etype == T_OBJECT);
-  int  tag   =  isobj ? _lh_array_tag_obj_value : _lh_array_tag_type_value;
+  int  tag   =  isobj ? _lh_array_tag_obj_value : (int)_lh_array_tag_type_value;
   int lh = array_layout_helper(tag, hsize, etype, exact_log2(esize));
 
   assert(lh < (int)_lh_neutral_value, "must look like an array layout");
